Upgraded to .NET 5 and separated the UI in the main 'Sanford.Multimedia.Midi' assembly into it's own assembly called 'Sanford.Multimedia.Midi.UI.Windows', while the rest of the Sanford.Multimedia.Midi code is now renamed to 'Sanford.Multimedia.Midi.Core', this will ensure cross-platform compatibility.
