+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\Sanford.Multimedia.Midi.sln",
+  "PreviewInSolutionExplorer": false
+}
