+++ Demo/SequencerDemo/SequencerDemo.Windows.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net5.0-windows</TargetFramework>
+    <OutputType>WinExe</OutputType>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+    <UseWindowsForms>true</UseWindowsForms>
+    <ImportWindowsDesktopTargets>true</ImportWindowsDesktopTargets>
+    <AssemblyName>SequencerDemo.Windows</AssemblyName>
+    <RootNamespace>SequencerDemo.Windows</RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Sanford.Multimedia.Midi.UI.Windows\Sanford.Multimedia.Midi.UI.Windows.csproj" />
+    <ProjectReference Include="..\..\Source\Sanford.Multimedia.Midi.Core.csproj" />
+  </ItemGroup>
+</Project>
