+++ Sanford.Multimedia.Midi.sln
@@ -1,15 +1,18 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SequencerDemo", "Demo\SequencerDemo\SequencerDemo.csproj", "{90480CE5-E392-4713-AB5C-7D659ACA0B85}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SequencerDemo.Windows", "Demo\SequencerDemo\SequencerDemo.Windows.csproj", "{90480CE5-E392-4713-AB5C-7D659ACA0B85}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MidiWatcher", "Demo\MidiWatcher\MidiWatcher.csproj", "{752C27BA-E092-455F-9B3F-1221E13FAE40}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MidiWatcher.Windows", "Demo\MidiWatcher\MidiWatcher.Windows.csproj", "{752C27BA-E092-455F-9B3F-1221E13FAE40}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sanford.Multimedia.Midi", "Source\Sanford.Multimedia.Midi.csproj", "{4269C72A-8D3A-4737-8F89-72EAA33EA9E1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sanford.Multimedia.Midi.Core", "Source\Sanford.Multimedia.Midi.Core.csproj", "{4269C72A-8D3A-4737-8F89-72EAA33EA9E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TimerTester", "TimerTester\TimerTester.csproj", "{E1F82BBB-3654-4615-8E6D-5A3D9DC308BC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sanford.Multimedia.Midi.UI.Windows", "Sanford.Multimedia.Midi.UI.Windows\Sanford.Multimedia.Midi.UI.Windows.csproj", "{F908E146-FD17-4853-B6C3-6AEAC148B420}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4269C72A-8D3A-4737-8F89-72EAA33EA9E1} = {4269C72A-8D3A-4737-8F89-72EAA33EA9E1}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,12 +32,15 @@ Global
 		{4269C72A-8D3A-4737-8F89-72EAA33EA9E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4269C72A-8D3A-4737-8F89-72EAA33EA9E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4269C72A-8D3A-4737-8F89-72EAA33EA9E1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E1F82BBB-3654-4615-8E6D-5A3D9DC308BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1F82BBB-3654-4615-8E6D-5A3D9DC308BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1F82BBB-3654-4615-8E6D-5A3D9DC308BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1F82BBB-3654-4615-8E6D-5A3D9DC308BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F908E146-FD17-4853-B6C3-6AEAC148B420}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F908E146-FD17-4853-B6C3-6AEAC148B420}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F908E146-FD17-4853-B6C3-6AEAC148B420}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F908E146-FD17-4853-B6C3-6AEAC148B420}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A0F89C5C-2D39-4686-8962-8311987EF148}
 	EndGlobalSection
 EndGlobal
